+++ climaf/operators.py
@@ -399,7 +399,7 @@ def fixed_fields(operator, *paths):
         ... ('mesh_zgr.nc','/data/climaf/${project}/${model}/ORCA1_mesh_zgr.nc'))
 
        >>> fixed_fields('plot',
-        ... ('coordinates.nc','/cnrm/aster/data3/aster/chevalli/Partage/NEMO/eORCA_R025_coordinates_v1.0.nc'))
+        ... ('coordinates.nc','/cnrm/ioga/Users/chevallier/chevalli/Partage/NEMO/eORCA_R025_coordinates_v1.0.nc'))
 
     """
     if not isinstance(operator,list):